opencl: keep the int8 attention P block scale in f32
The u8 P of the int8 KQV GEMM carried a half scale per 32-block, exp(blockmax - rowmax)/255.
When the row max sits far above the scores, which an attention sink of +8 or more does, that
scale falls below the half normal range. A device that flushes half subnormals then drops
every block of the row: on the X1-85 the FLASH_ATTN_EXT cases with sinks fail by up to 30%
NMSE with the int8 KQV on, while the same cases pass on the X2-90, where the subnormal
survives with reduced precision. f32 scales cost 2 bytes more per 32 scores of P.

(cherry picked from commit c28c1446fdad52485bed3adf8dd516eb2e616e9d)

+++ ggml/src/ggml-opencl/kernels/softmax_4_f16.cl
@@ -248,7 +248,7 @@ REQD_SUBGROUP_SIZE_64
 kernel void kernel_soft_max_4_f16_q8(
         global uchar * qp,
         ulong offset_qp,
-        global half * dp,
+        global float * dp,
         ulong offset_dp,
         int qp_pitch,               // bytes per P row, >= ne00, multiple of 32
         global float * psums,
@@ -324,7 +324,7 @@ kernel void kernel_soft_max_4_f16_q8(
     const uint row  = (uint)(i03*get_num_groups(1)*get_num_groups(0) + i02*get_num_groups(0) + i01);
 
     global uchar * qrow = (global uchar *)((global char *)qp + offset_qp) + (size_t)row*qp_pitch;
-    global half  * drow = (global half  *)((global char *)dp + offset_dp) + (size_t)row*(qp_pitch/32);
+    global float * drow = (global float *)((global char *)dp + offset_dp) + (size_t)row*(qp_pitch/32);
 
     float lsum = 0.0f;
     for (int b = get_local_id(0); b < nblk; b += get_local_size(0)) {
@@ -338,11 +338,14 @@ kernel void kernel_soft_max_4_f16_q8(
             amax = fmax(amax, fmax(fmax(e.s0, e.s1), fmax(e.s2, e.s3)));
         }
 
-        // P >= 0, so quantise unsigned: 255 levels instead of 127
+        // P >= 0, so quantise unsigned: 255 levels instead of 127. The scale stays f32: it is
+        // exp(blockmax - rowmax)/255, and a row whose max sits far above its scores (a large
+        // attention sink) puts it below the half normal range, where a device that flushes
+        // subnormals drops the whole block.
         const float d  = amax / 255.0f;
         const float id = amax > 0.0f ? 255.0f / amax : 0.0f;
 
-        drow[b] = (half)d;
+        drow[b] = d;
 
         #pragma unroll
         for (int i = 0; i < 32; ++i) {
@@ -366,7 +369,7 @@ kernel void kernel_soft_max_4_f16_q8(
 
 // Second half of the fused KQ+softmax path (mul_mm_f16_f32_kq_p8). One workgroup
 // per (query, head) row: reduce the per-32-block maxima to the row max (sinks join
-// it as an extra column), then emit the per-block half scale
+// it as an extra column), then emit the per-block f32 scale
 // exp(bmax - rowmax)/255 that kernel_mul_mm_q8_kqv consumes and the deferred-norm
 // row sum. Touches 8 bytes per 32 scores instead of the 13 bytes per score the
 // two-pass softmax read, so it is ~1/50 of that kernel's traffic.
@@ -376,7 +379,7 @@ REQD_SUBGROUP_SIZE_64
 kernel void kernel_fa_p8_fixup(
         global float * bmax,
         global float * bsum,
-        global half  * dp,
+        global float * dp,
         global float * psums,
         global char  * sinks,
         ulong offset_sinks,
@@ -390,7 +393,7 @@ kernel void kernel_fa_p8_fixup(
     const uint row = (uint)head*(uint)N + (uint)n;
     global float * bm = bmax + (size_t)row*row_blk;
     global float * bs = bsum + (size_t)row*row_blk;
-    global half  * dr = dp   + (size_t)row*row_blk;
+    global float * dr = dp   + (size_t)row*row_blk;
 
     float lmax = -INFINITY;
     for (int b = get_local_id(0); b < nblk; b += get_local_size(0)) {
@@ -409,7 +412,7 @@ kernel void kernel_fa_p8_fixup(
         // finite sentinels rather than -INFINITY compares: the program builds with
         // -cl-finite-math-only, which may fold a test against infinity itself.
         const float f = (m > -1.0e30f && rowmax > -1.0e30f) ? exp(m - rowmax) : 0.0f;
-        dr[b] = (half)(f / 255.0f);
+        dr[b] = f / 255.0f;   // f32: see kernel_soft_max_4_f16_q8
         lsum += bs[b]*f;
     }
     float sum = sub_group_reduce_add(lsum);
